dracut-android-userdata (#10)
properly packaged and written script that handles bootstrapping the
nested partition table residing in the Android userdata partition.

the hook handles doing the partition resize and mapping the nested
partitions with kpartx.

notably, this hook will *not* do the rootfs resize. instead, I'm going
to ship that as a small standalone systemd unit that relies on
ConditionFirstBoot=yes