image: include linux-firmware-qcom
turns out, dracut is already able to figure out that adreno module needs
adreno firmware.

the only bummer is it's including *all* Adreno firmwares (a3xx->a6xx)

it's a shame that ALARM isn't packaging compressed firmware like Arch
is. but as it stands it's only adding a mb or so to the initrd.

closes #9