use systemd-repart
since lk2nd doesn't support GPT, and systemd-repart doesn't support MBR,
just hack the GPT table to MBR as a post-processing step before
flashing. sgdisk makes it quite simple and it boots okay.

filesystem UUIDs are still hardcoded for now, not ideal but
systemd-repart specifically has support for deterministic filesystem
UUIDS (which are derived from part UUID + a seed that is random by
default but can be pinned as I have done in the mkosi.conf)